Rename flask module to web
