+++ localtileserver/__main__.py
@@ -1,5 +1,5 @@
 # Import as run_app for entry_point
-from localtileserver.tileserver.application import click_run_app as run_app
+from localtileserver.web.application import click_run_app as run_app
 
 if __name__ == "__main__":  # pragma: no cover
     run_app()
