--- localtileserver/tileserver/__init__.py
@@ -1,4 +1,0 @@
-# flake8: noqa: F401
-from localtileserver.tileserver import rest, urls, views
-from localtileserver.tileserver.application import create_app, run_app
-from localtileserver.tileserver.blueprint import cache, tileserver
