+++ localtileserver/web/static/js/cesium.js
