+++ localtileserver/web/static/js/geojsControls.js
