+++ localtileserver/web/templates/tileserver/_include/cesium.html
