+++ localtileserver/web/templates/tileserver/_include/examples.html
