+++ localtileserver/web/templates/tileserver/_include/geojs.html
