+++ localtileserver/web/templates/tileserver/_include/geojsControls.html
