+++ localtileserver/web/templates/tileserver/_include/palettes.html
