+++ localtileserver/web/templates/tileserver/base.html
