+++ localtileserver/web/templates/tileserver/baseTileViewer.html
