+++ localtileserver/web/templates/tileserver/cesiumSplitViewer.html
