+++ localtileserver/web/templates/tileserver/cesiumViewer.html
