+++ localtileserver/web/templates/tileserver/geojsViewer.html
