+++ localtileserver/web/templates/tileserver/splitForm.html
